Finalise matching process
